Created a third scene by duplicating 'Scene_2_Basic_Collectible', in this new scene (Scene_3_Basic_Collectible), the camera is a child of the Player, the Player object's z rotation is locked, the camera is larger, there is at least one more wall, the original wall is resized, and the camera is centered on the player.
Revamped the script 'Keyboard Movement' to stop the player when there is no key pressed while still supporting movement in 8 directions utilizing rb.velocity (replaces rb.AddForce method)

Note: There is still a problem with input delay. Try to fix this at some point, but for now movement is more functional than it was, if a bit uncomfortable.