Created new script attached to Meanie, 'MeanieMovesPlayer.cs'. Added functionality to script, moves a public GameObject (set to the Player gameObject), to global pos (0,0,0) which is where the player starts whenever the player collides with the/a meanie.
Note: possible issue with script, it utilizes OnTriggerEnter2D(Collider2D collider), will this trigger when meanie collides with something that isnt the player? I hope not. It doesnt trigger when the meanie runs into a wall or runs into a collectible, I tested that, but somewhere down the line this could cause a problem, figure out if its an issue and fix it sooner rather than later.