Added new feature/functionality, checkpoints! There are now checkpoints (purple squares) and a checkpoint manager in both the tutorial and game level, added new scripts, edited existing scripts, changed level design (added checkpoints to game level).
How do checkpoints work?

Checkpoints communicate their position to the checkpoint manager when the player collides with them then destroy.

Meanies check what the checkpoint manager's current checkpoint is (by default is start of the game, (0.0f, 0.0f)) by updating every frame, when the player collides with a meanie, it now sends the player to that checkpoint position instead of always to the start of the game.

Checkpoint -> checkpointManager <- meanie